#316 [refactor] UserMeetingSchedule 패키지 수정
